did-web3j-sdk/src/main/java/com/reddate/did/protocol/common/ClaimInfo.java
// Copyright 2021 Red Date Technology Ltd.  Licensed under MPLv2
// (https://www.mozilla.org/en-US/MPL/2.0/)
package com.reddate.did.protocol.common;

import java.io.Serializable;

public class ClaimInfo implements Serializable {

  private String name; // Field name

  private String type; // Field type, currently supports the basic data types of java

  private String desc; // Field description

  private Boolean
      required; // Whether the field is required, true is required, false is not required.

  public String getName() {
    return name;
  }

  public void setName(String name) {
    this.name = name;
  }

  public String getType() {
    return type;
  }

  public void setType(String type) {
    this.type = type;
  }

  public String getDesc() {
    return desc;
  }

  public void setDesc(String desc) {
    this.desc = desc;
  }

  public Boolean getRequired() {
    return required;
  }

  public void setRequired(Boolean required) {
    this.required = required;
  }
}
